promote VertexArray Layout template parameter from method to class
This migration went mostly smoothly, except for the total component count
calculation. I've left the existing calculation in for now because moving to a
calculation based on static Layout data requires a substantial change to how we
attach multiple vertex buffers to a vertex array.

It turns out I've misunderstood how multiple vertex buffers per vertex array
work. In the current implementation, we can only add multiple buffers of the
same layout. Multiple buffers are supposed to enable non-interleaved layouts,
where each buffer contains the data for one or more distinct attributes. Forcing
every vertex buffer to have the same layout prevents the use of, say, one buffer
for 3D positions and another for UV texture coordinates, which is a common use
case.

After turning VertexBuffer into a class template, I will try to change
VertexArray's single Layout template parameter into a parameter pack and replace
VertexArray::addVertexBuffer with a new method setVertexBuffers that takes a
tuple of buffers whose layouts match the parameter pack.

I will make that change after turning VertexBuffer into a class template,
because otherwise I don't think we can directly unroll the parameter pack in the
tuple field's declaration.

The main consequence of the change will be that the layouts of all vertex
buffers must be known when the vertex array is constructed, and all vertex
buffers must be constructed before any can be added to the vertex array. These
restrictions should work for all use cases in the forseeable future. If they
become a problem, it is possible to replace or supplement setVertexBuffers with
a new version of addVertexBuffer that produces a whole new VertexArray object
that extends the vertex array's template parameters.