drop static Attribute tagging
As the engine becomes more complex, it becomes harder to justify the cost of the
attribute tagging system. Since the vertex buffer layout is unlikely to require
this much flexibility any time soon, the extra generics and tupling machinery
add substantial complexity without much benefit, so I'm rolling it back.